fix: fix encrypt.txt filename to encrypted.txt
