+++ lib/encrypt.rb
@@ -1,4 +1,4 @@
 require_relative './enigma'
 enigma = Enigma.new
 
-enigma.read_message('./lib/message.txt', './lib/encrypt.txt', 'encrypt')
+enigma.read_message('./lib/message.txt', './lib/encrypted.txt', 'encrypt')
