--- lib/encrypt.txt
@@ -1,6 +1,0 @@
-keder ohulw!, 02715, 040895
-ivgqnmuvfyooethjefkzx, 04012, 040895
-fkqn, 35444, 060822
-keder ohulw!, 02715, 040895
-kuqypldchxywgsrrgeugz, 33920, 040895
-vdpd, 24707, 060822
